// frontend/src/app/player/shared/track.interface.ts
import {User} from './user.interface';

export interface Track {
  kind: String;
  id: Number;
  created_at: Date;
  user_id: Number;
  duration: Number;
  commentable: boolean;
  state: String;
  original_content_size: Number;
  last_modified: Date;
  sharing: String;
  tag_list: String;
  permalink: String;
  streamable: boolean;
  embeddable_by: String;
  downloadable: boolean;
  purchase_url: String;
  label_id: String;
  purchase_title: String;
  genre: String;
  title: String;
  description: String;
  label_name: String;
  release: String;
  track_type: String;
  key_signature: String;
  isrc: String;
  video_url: String;
  bpm: String;
  release_year: String;
  release_month: String;
  release_day: String;
  original_format: String;
  license: String;
  uri: URL;
  user: User;
  attachments_uri: URL;
  permalink_url: URL;
  artwork_url: String;
  waveform_url: URL;
  stream_url: URL;
  download_url: URL;
  playback_count: Number;
  download_count: Number;
  favoritings_count: Number;
  reposts_count: Number;
  comment_count: Number;
}